+++ src/gort/overwatcher/alerts.py
@@ -44,6 +44,7 @@ class AlertsSummary(BaseModel):
     o2_room_alerts: dict[str, bool] | None = None
     heater_alert: bool | None = None
     heater_camera_alerts: dict[str, bool] | None = None
+    engineering_override: bool = False
 
 
 class ConnectivityStatus:
@@ -69,10 +70,11 @@ class ActiveAlert(enum.Flag):
     DISCONNECTED = enum.auto()
     DOME_LOCKED = enum.auto()
     IDLE = enum.auto()
+    ENGINEERING_OVERRIDE = enum.auto()
     UNKNOWN = enum.auto()
 
     ALWAYS_CLOSE = HUMIDITY | DEW_POINT | WIND | RAIN
-    NO_CLOSE = DOOR | E_STOPS
+    NO_CLOSE = DOOR | E_STOPS | ENGINEERING_OVERRIDE
 
 
 class AlertsMonitorTask(OverwatcherModuleTask["AlertsOverwatcher"]):
@@ -101,6 +103,10 @@ class AlertsMonitorTask(OverwatcherModuleTask["AlertsOverwatcher"]):
                 n_failures = 0
             finally:
                 if self.module.unavailable is False and n_failures >= 5:
+                    await self.module.notify(
+                        "Alerts data is unavailable.",
+                        level="critical",
+                    )
                     self.module.unavailable = True
 
             await asyncio.sleep(self.INTERVAL)
@@ -227,6 +233,12 @@ class AlertsOverwatcher(OverwatcherModule):
             if self.idle_since > 0 and (time() - self.idle_since) > timeout:
                 self.log.warning(f"Overwatcher has been idle for over {timeout} s.")
                 active_alerts |= ActiveAlert.IDLE
+
+        # If the engineering mode is enabled, we assume it's safe.
+        if self.state.engineering_override:
+            self.log.warning("Engineering mode is enabled.")
+            active_alerts |= ActiveAlert.ENGINEERING_OVERRIDE
+            is_safe = True
 
         return is_safe, active_alerts