+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### 🔧 Fixed
 
-* Take into account the dither position when calculating `POSCIRA` and `POSCIDE` .
+* [#50](https://github.com/sdss/lvmgort/pull/50) Take into account the dither position when calculating `POSCIRA` and `POSCIDE` .
 * Fix logging the configuration in the pre-observing recipe.